+++ app/src/main/java/lk/test/myapplication/retrofit/task/TaskDto.java
@@ -1,4 +1,4 @@
-package lk.test.myapplication.models.task;
+package lk.test.myapplication.retrofit.task;
 
 // Task Data-transfer-object
 public class TaskDto {
