+++ app/src/main/res/layout/activity_login.xml
@@ -37,7 +37,8 @@
                 <EditText
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
-                    android:id="@+id/etPassword"/>
+                    android:id="@+id/etPassword"
+                    android:inputType="textPassword"/>
 
             </com.google.android.material.textfield.TextInputLayout>
